requirements.txt
discord.py
plex.py==0.9.0
PlexAPI==4.0.0
texttable
python-dotenv
jellyfin-apiclient-python
requests